+++ lib/deprecation_toolkit/minitest_hook.rb
@@ -7,7 +7,7 @@ module DeprecationToolkit
     def trigger_deprecation_toolkit_behavior
       current_deprecations = Collector.new(Collector.deprecations)
       recorded_deprecations = Collector.load(self)
-      if current_deprecations != recorded_deprecations
+      if !recorded_deprecations.flaky? && current_deprecations != recorded_deprecations
         Configuration.behavior.trigger(self, current_deprecations, recorded_deprecations)
       end
     ensure